// content-plus-workspace/projects/nucleus-lib/src/lib/enums/cookies.enum.ts
export enum CookiesEnum {
  appLang = 'healthy-plus-app-lang',
  Token = 'healthy-plus-token',
  RefreshToken = 'healthy-plus-refresh-token',
  UserProfile = 'healthy-plus-user-profile',
  StoredKeys = 'healthy-plus-stored-keys',
  OrderRequest = 'healthy-plus-order-request',
  DeviceId = 'healthy-plus-device-id',
  ActiveUserChat = 'healthy-plus-active-user-chat',
  ChatMissedMessage = 'healthy-plus-chat-missed-message',
  SignalrConnectionId = 'healthy-plus-signalr-connection-id',
  ChatUsers = 'healthy-plus-chat-users',
  BookingStatus = 'healthy-plus-booking-status',
}